Integrated esp_hidd_prf_api into hal_ble/hidd_le_prf_int
