+++ main/ble_hid/hal_ble.h
@@ -31,8 +31,6 @@
 #include "common.h"
 
 #include "esp_bt.h"
-
-#include "esp_hidd_prf_api.h"
 #include "esp_bt_defs.h"
 #include "esp_gap_ble_api.h"
 #include "esp_gatts_api.h"
