+++ main/ble_hid/hidd_le_prf_int.h
@@ -18,9 +18,91 @@
 #define __HID_DEVICE_LE_PRF__
 #include "esp_gatts_api.h"
 #include "esp_gatt_defs.h"
-#include "esp_hidd_prf_api.h"
 #include "esp_gap_ble_api.h"
 #include "hid_dev.h"
+
+
+typedef enum {
+    ESP_HIDD_EVENT_REG_FINISH = 0,                     
+    ESP_BAT_EVENT_REG,
+    ESP_HIDD_EVENT_DEINIT_FINISH,                       
+    ESP_HIDD_EVENT_BLE_CONNECT,                         
+    ESP_HIDD_EVENT_BLE_DISCONNECT,
+    ESP_HIDD_EVENT_BLE_VENDOR_REPORT_WRITE_EVT,
+} esp_hidd_cb_event_t;
+
+/// HID config status
+typedef enum {
+    ESP_HIDD_STA_CONN_SUCCESS = 0x00,
+    ESP_HIDD_STA_CONN_FAIL    = 0x01,
+} esp_hidd_sta_conn_state_t;
+
+/// HID init status
+typedef enum {
+    ESP_HIDD_INIT_OK = 0,
+    ESP_HIDD_INIT_FAILED = 1,
+} esp_hidd_init_state_t;
+
+/// HID deinit status
+typedef enum {
+    ESP_HIDD_DEINIT_OK = 0,
+    ESP_HIDD_DEINIT_FAILED = 0,
+} esp_hidd_deinit_state_t;
+
+/**
+ * @brief HIDD callback parameters union 
+ */
+typedef union {
+    /**
+	 * @brief ESP_HIDD_EVENT_INIT_FINISH
+	 */
+    struct hidd_init_finish_evt_param {
+        esp_hidd_init_state_t state;				/*!< Initial status */
+        esp_gatt_if_t gatts_if;
+    } init_finish;							      /*!< HID callback param of ESP_HIDD_EVENT_INIT_FINISH */
+
+    /**
+	 * @brief ESP_HIDD_EVENT_DEINIT_FINISH
+	 */
+    struct hidd_deinit_finish_evt_param {
+        esp_hidd_deinit_state_t state;				/*!< De-initial status */
+    } deinit_finish;								/*!< HID callback param of ESP_HIDD_EVENT_DEINIT_FINISH */
+
+    /**
+     * @brief ESP_HIDD_EVENT_CONNECT
+	 */
+    struct hidd_connect_evt_param {
+        uint16_t conn_id;
+        esp_bd_addr_t remote_bda;                   /*!< HID Remote bluetooth connection index */
+    } connect;									    /*!< HID callback param of ESP_HIDD_EVENT_CONNECT */
+
+    /**
+     * @brief ESP_HIDD_EVENT_DISCONNECT
+	 */
+    struct hidd_disconnect_evt_param {
+        esp_bd_addr_t remote_bda;                   /*!< HID Remote bluetooth device address */
+    } disconnect;									/*!< HID callback param of ESP_HIDD_EVENT_DISCONNECT */
+
+    /**
+     * @brief ESP_HIDD_EVENT_BLE_VENDOR_REPORT_WRITE_EVT
+	 */
+    struct hidd_vendor_write_evt_param {
+        uint16_t conn_id;                           /*!< HID connection index */
+        uint16_t report_id;                         /*!< HID report index */
+        uint16_t length;                            /*!< data length */
+        uint8_t  *data;                             /*!< The pointer to the data */
+    } vendor_write;									/*!< HID callback param of ESP_HIDD_EVENT_BLE_VENDOR_REPORT_WRITE_EVT */
+
+} esp_hidd_cb_param_t;
+
+
+/**
+ * @brief HID device event callback function type
+ * @param event : Event type
+ * @param param : Point to callback parameter, currently is union type
+ */
+typedef void (*esp_hidd_event_cb_t) (esp_hidd_cb_event_t event, esp_hidd_cb_param_t *param);
+
 
 //HID BLE profile log tag
 #define HID_LE_PRF_TAG                        "HID_LE_PRF"
